backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
Pillow==11.1.0
ultralytics==8.3.57
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pydantic>=2.0.0
python-dotenv==1.0.1
